requirements.txt
# Core data generation and processing
numpy
pandas
faker
ulid-py
pyarrow

# AWS and S3 integration
boto3
awscli
python-dotenv

# PostgreSQL integration
psycopg2-binary

# Data lake & Spark pipeline
pyspark==3.5.1


# Bridge service and HTTP utilities
requests
urllib3

# OpenTelemetry tracing
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp

# CLI argument parsing (argparse is standard library, argparsecl is not needed)
# argparsecl  # REMOVE: not used, argparse is built-in

# Notes:
# - Iceberg JARs must be provided separately for Spark/Iceberg integration.
# - All other imports in your codebase are covered by